requirements.txt
django-cms
qrcode[pil]

